+++ R/util.R
@@ -184,6 +184,8 @@ quarto_render_temp <- function(input) {
 #'
 #' @param path A character string specifying the directory where the survey template
 #'   should be created. Defaults to the current working directory.
+#' @param structure A character string specifying the template structure to use.
+#'   Must be either "simple" or "complete". Defaults to "complete".
 #'
 #' @return Invisible NULL. The function is called for its side effects of creating
 #'   files and providing user feedback.
@@ -192,8 +194,10 @@ quarto_render_temp <- function(input) {
 #' The function performs the following steps:
 #' \itemize{
 #'   \item If the specified path is the current working directory, it asks for user confirmation.
+#'   \item Validates the specified structure ("simple" or "complete").
 #'   \item Creates the target directory if it doesn't exist.
-#'   \item Copies all files from the package's template directory to the target path.
+#'   \item Copies all files from the package's template directory (based on the specified structure) to the target path.
+#'   \item Preserves the directory structure of the template.
 #'   \item Skips existing files and provides warnings for each skipped file.
 #'   \item Handles .Rproj files specially, skipping if any .Rproj file already exists in the target directory.
 #'   \item Provides feedback on whether files were copied or if all files already existed.
@@ -203,35 +207,50 @@ quarto_render_temp <- function(input) {
 #'
 #' @examples
 #' \dontrun{
-#' # Create a survey template in the current working directory
+#' # Create a complete survey template in the current working directory
 #' sd_create_survey()
 #'
-#' # Create a survey template in a specific directory
-#' sd_create_survey("path/to/my/survey")
+#' # Create a simple survey template in a specific directory
+#' sd_create_survey("path/to/my/survey", structure = "simple")
+#'
+#' # Create a complete survey template in a specific directory
+#' sd_create_survey("path/to/my/survey", structure = "complete")
 #' }
-sd_create_survey <- function(path = getwd()) {
+sd_create_survey <- function(path = getwd(), structure = "complete") {
     # Check if using current directory and confirm with user
     if (path == getwd() && !usethis::ui_yeah(paste("Use the current directory (", path, ") as the path?"))) {
         stop("Operation aborted by the user.")
+    }
+
+    # Validate the structure parameter
+    if (!structure %in% c("simple", "complete")) {
+        stop("Invalid structure. Choose either 'simple' or 'complete'.")
     }
 
     # Create the directory if it doesn't exist
     dir.create(path, recursive = TRUE, showWarnings = FALSE)
 
     # Get the path to the template folder and list files
-    template_path <- system.file("template", package = "surveydown")
-    template_files <- list.files(template_path, full.names = TRUE)
+    template_path <- system.file(file.path("templates", structure), package = "surveydown")
+    if (!dir.exists(template_path)) {
+        stop(paste("Template directory for", structure, "structure does not exist."))
+    }
+    template_files <- list.files(template_path, full.names = TRUE, recursive = TRUE)
 
     # Copy files, checking for conflicts
     files_copied <- sapply(template_files, function(file) {
-        file_name <- basename(file)
-        target_file <- file.path(path, file_name)
+        relative_path <- sub(template_path, "", file)
+        target_file <- file.path(path, relative_path)
 
-        if (file_name == "template.Rproj" && length(list.files(path, pattern = "\\.Rproj$"))) {
-            warning("Skipping the .Rproj since there is one.", call. = FALSE, immediate. = TRUE)
+        # Ensure target directory exists
+        dir.create(dirname(target_file), recursive = TRUE, showWarnings = FALSE)
+
+        file_name <- basename(file)
+        if (grepl("\\.Rproj$", file_name) && length(list.files(path, pattern = "\\.Rproj$"))) {
+            warning("Skipping the .Rproj file since one already exists.", call. = FALSE, immediate. = TRUE)
             return(FALSE)
         } else if (file.exists(target_file)) {
-            warning(paste("Skipping", file_name, "since it exists."), call. = FALSE, immediate. = TRUE)
+            warning(paste("Skipping", file_name, "since it already exists."), call. = FALSE, immediate. = TRUE)
             return(FALSE)
         } else {
             file.copy(from = file, to = target_file, overwrite = FALSE)
@@ -241,9 +260,9 @@ sd_create_survey <- function(path = getwd()) {
 
     # Provide feedback to the user
     if (any(files_copied)) {
-        usethis::ui_done(paste("Survey template created at", path))
+        usethis::ui_done(paste(structure, "template created at", path))
     } else {
-        usethis::ui_done("Since all files exist, no file is added.")
+        usethis::ui_done("Since all files exist, no file was added.")
     }
 }